Rename package name to com.globe
The domain is not available, but this project isn't meant to be an actual product either.
