+++ src/main/java/com/globe/Globe.java
@@ -1,0 +1,12 @@
+package com.globe;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class Globe {
+
+    public static void main(String[] args) {
+        SpringApplication.run(Globe.class, args);
+    }
+}
