+++ src/main/java/com/globe/city/CityNotFoundException.java
@@ -1,0 +1,10 @@
+package com.globe.city;
+
+import com.globe.common.ResourceNotFoundException;
+
+public class CityNotFoundException extends ResourceNotFoundException {
+
+    public CityNotFoundException() {
+        super();
+    }
+}
