+++ src/main/java/com/globe/common/ResourceNotFoundException.java
@@ -1,0 +1,12 @@
+package com.globe.common;
+
+import org.springframework.http.HttpStatus;
+import org.springframework.web.bind.annotation.ResponseStatus;
+
+@ResponseStatus(value = HttpStatus.NOT_FOUND)
+public class ResourceNotFoundException extends RuntimeException {
+
+    public ResourceNotFoundException() {
+        super();
+    }
+}
