+++ src/main/java/com/globe/country/CountryNotFoundException.java
@@ -1,0 +1,10 @@
+package com.globe.country;
+
+import com.globe.common.ResourceNotFoundException;
+
+public class CountryNotFoundException extends ResourceNotFoundException {
+
+    public CountryNotFoundException() {
+        super();
+    }
+}
