+++ src/main/java/com/globe/country/NoSuchContinentException.java
@@ -1,0 +1,10 @@
+package com.globe.country;
+
+import java.util.NoSuchElementException;
+
+public class NoSuchContinentException extends NoSuchElementException {
+
+    public NoSuchContinentException(String message) {
+        super(message);
+    }
+}
