+++ src/main/java/com/globe/language/LanguageNotFoundException.java
@@ -1,0 +1,10 @@
+package com.globe.language;
+
+import com.globe.common.ResourceNotFoundException;
+
+public class LanguageNotFoundException extends ResourceNotFoundException {
+
+    public LanguageNotFoundException() {
+        super();
+    }
+}
