--- src/main/java/io/github/callmeneva/globe/Globe.java
@@ -1,12 +1,0 @@
-package io.github.callmeneva.globe;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class Globe {
-
-    public static void main(String[] args) {
-        SpringApplication.run(Globe.class, args);
-    }
-}
