--- src/main/java/io/github/callmeneva/globe/ModelMapperConfig.java
@@ -1,14 +1,0 @@
-package io.github.callmeneva.globe;
-
-import org.modelmapper.ModelMapper;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-
-@Configuration
-public class ModelMapperConfig {
-
-    @Bean
-    public ModelMapper modelMapper() {
-        return new ModelMapper();
-    }
-}
