--- src/main/java/io/github/callmeneva/globe/city/CityNotFoundException.java
@@ -1,10 +1,0 @@
-package io.github.callmeneva.globe.city;
-
-import io.github.callmeneva.globe.common.ResourceNotFoundException;
-
-public class CityNotFoundException extends ResourceNotFoundException {
-
-    public CityNotFoundException() {
-        super();
-    }
-}
