--- src/main/java/io/github/callmeneva/globe/common/ResourceNotFoundException.java
@@ -1,12 +1,0 @@
-package io.github.callmeneva.globe.common;
-
-import org.springframework.http.HttpStatus;
-import org.springframework.web.bind.annotation.ResponseStatus;
-
-@ResponseStatus(value = HttpStatus.NOT_FOUND)
-public class ResourceNotFoundException extends RuntimeException {
-
-    public ResourceNotFoundException() {
-        super();
-    }
-}
