--- src/main/java/io/github/callmeneva/globe/country/CountryNotFoundException.java
@@ -1,10 +1,0 @@
-package io.github.callmeneva.globe.country;
-
-import io.github.callmeneva.globe.common.ResourceNotFoundException;
-
-public class CountryNotFoundException extends ResourceNotFoundException {
-
-    public CountryNotFoundException() {
-        super();
-    }
-}
