--- src/main/java/io/github/callmeneva/globe/country/NoSuchContinentException.java
@@ -1,10 +1,0 @@
-package io.github.callmeneva.globe.country;
-
-import java.util.NoSuchElementException;
-
-public class NoSuchContinentException extends NoSuchElementException {
-
-    public NoSuchContinentException(String message) {
-        super(message);
-    }
-}
