--- src/main/java/io/github/callmeneva/globe/language/LanguageNotFoundException.java
@@ -1,10 +1,0 @@
-package io.github.callmeneva.globe.language;
-
-import io.github.callmeneva.globe.common.ResourceNotFoundException;
-
-public class LanguageNotFoundException extends ResourceNotFoundException {
-
-    public LanguageNotFoundException() {
-        super();
-    }
-}
